# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinZ LMS API",
        "description": "API documentation for FinZ LMS backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/compliance/returns": {
            "post": {
                "tags": [
                    "ComplianceReport"
                ],
                "summary": "generateReturn ComplianceReport",
                "operationId": "381995c41d4cc0b828e8eaf2924694ae",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/compliance/dpdp-requests": {
            "get": {
                "tags": [
                    "ComplianceReport"
                ],
                "summary": "dpdpRequests ComplianceReport",
                "operationId": "09b5a3b0db4d7179c9f6b1c1b37615c5",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/compliance/dpdp-requests/{id}/resolve": {
            "post": {
                "tags": [
                    "ComplianceReport"
                ],
                "summary": "resolveDpdpRequest ComplianceReport",
                "operationId": "1434cd0edb54aa7abf49343c7466c8f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/compliance/data-masking-policy": {
            "get": {
                "tags": [
                    "ComplianceReport"
                ],
                "summary": "dataMaskingPolicy ComplianceReport",
                "operationId": "fd1424eff5edc0c3086ccf2b94ec76d0",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ComplianceReport"
                ],
                "summary": "updateDataMaskingPolicy ComplianceReport",
                "operationId": "24d09628819b1026ab6b274f5561133e",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/compliance/retention-policy": {
            "get": {
                "tags": [
                    "ComplianceReport"
                ],
                "summary": "retentionPolicy ComplianceReport",
                "operationId": "2fc264a4ee5b5ca6e365ae18beef6f44",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ComplianceReport"
                ],
                "summary": "updateRetentionPolicy ComplianceReport",
                "operationId": "35dd66a32b7feeb775dc2956e70b95cb",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/compliance/dashboard": {
            "get": {
                "tags": [
                    "ComplianceReport"
                ],
                "summary": "dashboard ComplianceReport",
                "operationId": "4d173736385ee2ce20328be35945070d",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/consents": {
            "get": {
                "tags": [
                    "ConsentLog"
                ],
                "summary": "index ConsentLog",
                "operationId": "dfb35fd29f943267e7942f150544106d",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/consents/{id}/withdraw": {
            "post": {
                "tags": [
                    "ConsentLog"
                ],
                "summary": "withdraw ConsentLog",
                "operationId": "da927bd851ee4347507583193b1f2f48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/consents/{id}/diff/{compare_id}": {
            "get": {
                "tags": [
                    "ConsentLog"
                ],
                "summary": "diff ConsentLog",
                "operationId": "43ce7f6c5c4c739dce16cc0171b5b0c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "compare_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/consents/export": {
            "get": {
                "tags": [
                    "ConsentLog"
                ],
                "summary": "export ConsentLog",
                "operationId": "fb04d75701ff02f1d2cb5d39bfa30eb9",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}